test(pg): the embedding repair test requires the test database instead of skipping (#878)
src/embedding-repair.pg.test.ts read OPENBRAIN_TEST_DATABASE_URL itself and
swapped in describe.skip when it was unset, so a database-less run reported
0 pass / 18 skip / 0 fail and exited 0 -- indistinguishable at the exit code
from a suite that ran and passed.

The file now calls requireTestDatabaseUrl() at module scope: absent the
variable the run throws test_database_required and fails loudly.

The suite is split by subject so each half stays under the lint standard the
pre-commit gate applies to a staged file whole:

- src/embedding-repair.pg.test.ts keeps the six scope-and-drift tests as one
  plain describe("embedding repair scope and drift (live Postgres)").
- src/embedding-repair-formula.pg.test.ts holds the four writer-parity suites
  (decisions, decision tag merge, sessions, real maintenance queue).
- src/embedding-repair-test-helpers.ts holds the seeding, cleanup, and
  inspection helpers both halves share. It holds no test and creates no pool:
  every function that touches the database takes the caller's pool as its
  first parameter, so each file owns and ends exactly one connection.

Test names, order, and assertions are unchanged. Each it body is hoisted to a
module-scope async function so no describe callback exceeds the per-function
line standard, and every non-null assertion is replaced by an expectDefined
guard exported from the helper module.

scripts/assert-db-tests-ran.ts registers all five describe names at the counts
JUnit emits (6 + 3 + 2 + 3 + 2) and raises MIN_TOTAL_LIVE_TESTCASES from 285
to 301, so CI's anti-skip guard fails if any of them stops running.

Verified:
- CHANGED_FILES="src/embedding-repair.pg.test.ts" bash
  scripts/done-means/878-pg-tests-require-database.sh -> exit 1 at origin/main
  (clauses 1-3 FAIL).
- CHANGED_FILES="src/embedding-repair.pg.test.ts src/embedding-repair-formula.pg.test.ts"
  bash scripts/done-means/878-pg-tests-require-database.sh -> exit 0, all four
  clauses PASS.
- bun run test:isolated over both files -> exit 0, 16 tests across 2 files.
- oxlint --deny-warnings on all four files -> exit 0; bunx tsc --noEmit -> exit 0.
